directory renames pathing changes
