+++ index.php
@@ -8,8 +8,8 @@
  * @copyright Copyright (c) 2018-2019 Jerry Padgett <sjpadgett@gmail.com>
  * @license   https://github.com/openemr/openemr/blob/master/LICENSE GNU General Public License 3
  */
-require_once($GLOBALS['rootdir'] . "globals.php");
+require_once(__DIR__ . "/../../../globals.php");
 
-use OpenEMR\Modules\oeFax\Controller\AppDispatch;
+use OpenEMR\Modules\FaxSMS\Controllers\AppDispatch;
 
 AppDispatch::getApiService();
