+++ rc_sms_notification.php
@@ -10,7 +10,7 @@
 // Rework of original
 ////////////////////////////////////////////////////////////////////
 //hack add for command line version
-use OpenEMR\Modules\oeFax\Controller\AppDispatch;
+use OpenEMR\Modules\FaxSMS\Controllers\AppDispatch;
 
 $_SERVER['REQUEST_URI'] = $_SERVER['PHP_SELF'];
 $_SERVER['SERVER_NAME'] = 'localhost';
@@ -27,7 +27,7 @@ if ($argc > 1 && empty($_SESSION['site_id']) && empty($_GET['site'])) {
     $_GET['site'] = isset($args[1]) ? $args[1] : 'default';
 }
 
-require_once($GLOBALS['rootdir'] . "globals.php");
+require_once("../../interface/globals.php");
 require_once("$srcdir/appointments.inc.php");
 
 $clientApp = AppDispatch::getApiService();
